// src/app/@core/models/policy.ts
export enum Policy {
  AdminCreate = 'admin.create',
  AdminDelete = 'admin.delete',
  AdminInvite = 'admin.invite',
  AdminRead = 'admin.read',
  AdminRemoteStart = 'admin.remote-start',
  AdminRemoteStop = 'admin.remote-stop',
  AdminReset = 'admin.reset',
  AdminUnlock = 'admin.unlock',
  AdminUpdate = 'admin.update',
  CardCreate = 'card.create',
  CardDelete = 'card.delete',
  CardRead = 'card.read',
  CardUpdate = 'card.update',
  ChargePointCreate = 'charge-point.create',
  ChargePointRead = 'charge-point.read',
  ChargePointRemoteStart = 'charge-point.remote-start',
  ChargePointRemoteStop = 'charge-point.remote-stop',
  ChargePointReset = 'charge-point.remote-reset',
  ChargePointUnlock = 'charge-point.remote-unlock',
  ChargePointUpdate = 'charge-point.update',
  ChargePointDelete = 'charge-point.delete',
  ChargePointPhotoUpload = 'charge-point.photo-upload',
  ChargePointPhotoDelete = 'charge-point.photo-delete',
  ChargeSessionRead = 'charge-session.read',
  CampaignCreate = 'campaign.create',
  CampaignUpdate = 'campaign.update',
  CampaignDelete = 'campaign.delete',
  CampaignRead = 'campaign.read',
  CampaignPhotoUpload = 'campaign.photo-upload',
  CampaignPhotoDelete = 'campaign.photo-delete',
  CouponCreate = 'coupon.create',
  CouponUpdate = 'coupon.update',
  CouponDelete = 'coupon.delete',
  CouponRead = 'coupon.read',
  CityUpdate = 'city.update',
  CountryRead = 'country.read',
  ConnectorManage = 'connector.manage',
  ConnectorUpdate = 'connector.update',
  CreditCardCreate = 'credit-card.create',
  CreditCardDelete = 'credit-card.delete',
  CreditCardRead = 'credit-card.read',
  CreditCardUpdate = 'credit-card.update',
  DeviceTokenPush = 'device-token.push',
  DirectoryRead = 'directory.read',
  DirectorySwitch = 'directory.switch',
  DirectoryCreate = 'directory.create',
  DirectoryUpdate = 'directory.update',
  DirectoryDelete = 'directory.delete',
  DistrictRead = 'district.read',
  SiteHostInvoiceRequestCreate = 'site-host-invoice-request.create',
  InvoiceRequestCreate = 'invoice-request.create',
  InvoiceRequestUpdate = 'invoice-request.update',
  InvoiceRequestDelete = 'invoice-request.delete',
  InvoiceRequestRead = 'invoice-request.read',
  InvoiceRequestPhotoUpload = 'invoice-request.photo-upload',
  InvoiceRequestPhotoDelete = 'invoice-request.photo-delete',
  IntegrationCreate = 'directory-integration.create',
  IntegrationUpdate = 'directory-integration.update',
  IntegrationDelete = 'directory-integration.delete',
  IntegrationRead = 'directory-integration.read',
  PaymentMethodRead = 'payment-method.read',
  PolicyRead = 'policy.read',
  PolicyUpdate = 'policy.update',
  ReservationRead = 'reservation.read',
  ReviewCreate = 'review.create',
  ReviewRead = 'review.read',
  RoleRead = 'role.read',
  RoleDelete = 'role.delete',
  PaymentRead = 'subscription-payments.read',
  SiteHostCreate = 'site-host.create',
  SiteHostUpdate = 'site-host.update',
  SiteHostDelete = 'site-host.delete',
  SiteHostRead = 'site-host.read',
  TariffCreate = 'tariff.create',
  TariffDelete = 'tariff.delete',
  TariffRead = 'tariff.read',
  TariffUpdate = 'tariff.update',
  TransactionRead = 'transaction.read',
  UserInvite = 'user.invite',
  UserRead = 'user.read',
  UserDelete = 'user.delete',
  UserUpdate = 'user.update',
  WorkspaceCreate = 'workspace.create',
  WorkspaceDelete = 'workspace.delete',
  WorkspaceRead = 'workspace.read',
  WorkspaceUpdate = 'workspace.update',
}